requirements.txt
Faker==22.7.0
pandas==2.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
streamlit==1.31.0
boto3==1.34.35
python-dotenv==1.0.1
sqlparse==0.4.4